Session_2_4/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
black==25.1.0
blinker==1.9.0
boto3==1.40.22
botocore==1.40.22
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.0.9
click==8.2.1
cohere==5.17.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
courlan==1.3.2
coverage==7.10.6
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fasta2a==0.3.2
fastapi==0.116.1
fastapi-cli==0.0.7
fastavro==1.12.0
ffmpy==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.7.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.1
google-genai==1.32.0
googleapis-common-protos==1.70.0
gradio==5.35.0
gradio_client==1.10.4
graphiti==0.1.13
graphiti-core==0.12.4
griffe==1.13.0
groovy==0.1.2
groq==0.31.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
invoke==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
logfire-api==4.3.6
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.13.1
mdurl==0.1.2
mistralai==1.9.10
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.5.0
neo4j==5.28.1
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.79.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.4.0
pluggy==1.6.0
posthog==4.0.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.4
psutil==5.9.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.3.2
pydantic-ai-slim==0.3.2
pydantic-evals==0.3.2
pydantic-graph==0.3.2
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.1
Pygments==2.19.2
PyMuPDF==1.25.5
PyMySQL==1.1.2
pyparsing==3.2.3
PyPika==0.48.9
pyppeteer==2.0.0
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.0.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.25.0
rsa==4.9.1
ruff==0.12.11
s3transfer==0.13.1
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==2.3.6
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tld==0.13.1
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
typer==0.15.4
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
